while loop change
Parsing did not quite work as written in the original code
As my host and ip is in CH and the webpage https://isitchristmas.com/ is dynamic I first thought the reason was the title being changed to 'NEIN/NON'. This was not the case, for it still did not work after the change to 'NEIN/NON'. Therefore the simplification